pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["sky-sky-code <saidyllapashaev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
websockets = "^12.0"
tortoise-orm = "^0.20.0"
aerich = "^0.7.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"



[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
